Fixed id datatype

+++ src/storage/sql/models/profile.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, Integer, String, ForeignKey
+from sqlalchemy import BigInteger, Column, DateTime, Integer, String, ForeignKey
 from sqlalchemy.dialects.postgresql import ARRAY
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -13,7 +13,7 @@ from src.storage.sql.models import Base
 class Profile(Base):
     __tablename__ = "profiles"
 
-    id: Mapped[int] = mapped_column(Integer, ForeignKey('users.id'), primary_key=True)
+    id: Mapped[int] = mapped_column(BigInteger, ForeignKey('users.id'), primary_key=True)
     
     # Attributes
     date_of_birth: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=True)
